Friends_Unit_Tests
This is Chad George from your CS 261 course.  I tried to do a pull request, but for the life of me, could not get it to work.  Anyways, I have implemented unit tests for the friends feature.  The tests, test both the admin and normal user functionality and should cover just about everything desired functionality in the API Rubric.